Fragment pause functions, hint solve system
