+++ base-hack/asm/objects.asm
@@ -47,7 +47,6 @@
 .importobj "obj/src_misc_item_db.o"
 .importobj "obj/src_misc_krusha.o"
 .importobj "obj/src_misc_move_hints.o"
-.importobj "obj/src_misc_pause_menu.o"
 .importobj "obj/src_misc_perma_kong_loss.o"
 .importobj "obj/src_misc_remove_blockers.o"
 .importobj "obj/src_misc_shop_indicator.o"
@@ -56,6 +55,10 @@
 .importobj "obj/src_misc_tns_number.o"
 .importobj "obj/src_misc_warp_to_isles.o"
 .importobj "obj/src_misc_win_condition.o"
+.importobj "obj/src_pause_carousel.o"
+.importobj "obj/src_pause_hints.o"
+.importobj "obj/src_pause_lib.o"
+.importobj "obj/src_pause_regions.o"
 .importobj "obj/src_randomizers_boss_rando.o"
 .importobj "obj/src_randomizers_every_lz_rando.o"
 .importobj "obj/src_randomizers_kut_out.o"
