+++ base-hack/src/pause/pause.h
@@ -1,0 +1,2 @@
+extern char display_billboard_fix;
+extern char hints_initialized;
