+++ base-hack/test.json
@@ -179,7 +179,7 @@
    "short_bosses":1,
    "fast_warp":1,
    "activate_all_bananaports":1,
-   "progressive_hint_gb_cap":60,
+   "progressive_hint_gb_cap":1,
    "arcade_order":[
       2,
       3,
